mm, treewide: make get_free_pages() and friends return void *
Vast majority of allocations that use get_free_pages() and it's
derivatives cast the returned value to a pointer type.

Make get_free_pages() and it derivatives return 'void *' with the exception of __get_dma_pages() whose callers all expect an unsigned long.

The call sites are updated with the semantic patch below and a bit of
manual formatting tweaking.

    @@
    type T;
    T addr;
    expression gfp, order;
    @@
    (
    - addr = __get_free_pages(gfp, order)
    + addr = (T)__get_free_pages(gfp, order)
    |
    - addr = __get_free_page(gfp)
    + addr = (T)__get_free_page(gfp)
    |
    - addr = get_zeroed_page(gfp)
    + addr = (T)get_zeroed_page(gfp)
    )

    @@
    type T;
    @@
    (
    - (T *)
      \(__get_free_pages\|get_zeroed_page\|__get_free_page\)(...)
    )

+++ arch/arm64/kvm/arm.c
@@ -2571,7 +2571,8 @@ static int __init init_hyp_mode(void)
 	for_each_possible_cpu(cpu) {
 		unsigned long stack_base;
 
-		stack_base = __get_free_pages(GFP_KERNEL, NVHE_STACK_SHIFT - PAGE_SHIFT);
+		stack_base = (unsigned long)__get_free_pages(GFP_KERNEL,
+							     NVHE_STACK_SHIFT - PAGE_SHIFT);
 		if (!stack_base) {
 			err = -ENOMEM;
 			goto out_err;
