+++ arch/powerpc/kvm/book3s_64_mmu_hv.c
@@ -88,8 +88,8 @@ int kvmppc_allocate_hpt(struct kvm_hpt_info *info, u32 order)
 	}
 
 	if (!hpt)
-		hpt = __get_free_pages(GFP_KERNEL|__GFP_ZERO|__GFP_RETRY_MAYFAIL
-				       |__GFP_NOWARN, order - PAGE_SHIFT);
+		hpt = (unsigned long)__get_free_pages(GFP_KERNEL | __GFP_ZERO | __GFP_RETRY_MAYFAIL | __GFP_NOWARN,
+						      order - PAGE_SHIFT);
 
 	if (!hpt)
 		return -ENOMEM;
